Allow recent posts to be shown, via "snippets".
Snippets are plugins that just populate the global variables
we setup in #2, and allow shared blocks to be viewed on any
page.

Currently we just add them to the sidebar in our sample theme
as a proof of concept.

Chosen the prefix `Snippets` as these can then be disabled
en-mass.