Avoid duplication the /index.rss template.
Given that RSS is a standard there is no need to contain
multiple copies.  Instead we have a master-template which
is included inline in RSS.pm.

If a local theme does contain an index.rss template then
it will still be used, but otherwise we've avoided some
duplication.

This closes #26.