Correctly link to the full-entry.
When testing this in my own blog I discovered that all the links
were relative, because we just used the "link" attribute from the
blog-post, rather than qualifying it with the content of the
'top' configuration-option.

In our templates we resolve this, by and large, but because the
HTML is in our code - a mistake - we're forced to perform the
expansion directly, and manually.