Bugfix: changed the inplacepost behaviour for root input folder posts
This fix addresses the problem of using InPlacePosts with post that live
on the root of the input directory.

Before the file links would have a something like /about.html, so when
the file is generated it would be output//about.html.

In some cases the regexes would fail and the link would end up being:

apost.post/about.html

which is just not right.

This bug affected any file (blog post or static page) that was on the
input root.