fix archive and tags
The archives and tag indexes wouldn't be recreated if the respective
directories existed but not their contents.
Also, the way of obtaining localized month names relied on
undocumented variables in Date::Language::* and didn't work for the
archive index. This uses the official Date::Language API now.
The fact that we have to decode Latin-1 strings from Date::Language
for most languages will in turn cause problems with Greek, Russian/KOI8 and
similar but that's something to be fixed in Date::Language and as it is it will
work for most available languages.